Ignore redundant nesting when checking for related siblings
If the best candidate is in an element all by itself, then we should
probably check its nearest ancestor that has siblings when considering
whether to append siblings that meet the score threshold.

For example, in the example below, we would now include the second
paragraph whereas previous we would not.


```html
<div>
  <div>
    <p>This is the best candidate.</p>
  </div>
</div>
<p>This paragraph meets the score threshold.</p>
```